\name{helpPlot}
\alias{helpPlot}
\docType{package}
\title{
MIfuns: Plotting
}
\details{
I want to ....
\itemize{
\item \link[=NONR]{Run} NONMEM and create diagnostic plots
\item \link[=covplot]{Make} "forest plots" of covariate effects, etc.
\item \link[=nm]{Create} and manipulate 'nm' objects
\item \link[=nmPlots]{Characterize} 'nm' objects with diagnostic plots
\item \link[=plot.nm]{Plot} 'nm' objects
}
}
